It doesn't log out when there is no user

+++ app/Http/Controllers/ChatbotLogoutController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use BotMan\BotMan\BotMan;
+use Illuminate\Support\Facades\Auth;
 
 class ChatbotLogoutController extends Controller
 {
@@ -16,6 +17,12 @@ class ChatbotLogoutController extends Controller
      */
     public function __invoke(BotMan $bot)
     {
+        if (Auth::guest()) {
+            $bot->reply('There is no logged-in user.');
+
+            return;
+        }
+
         $userData = $bot->userStorage()->find();
 
         $bot->userStorage()->delete();
